Fix file parameter validation
File parameter validation fails silently when the path specified doesn't
reference a file on the client's system. This requirement doesn't make
any sense, as it is not relevant whether the file exists on the client's
system, only whether it exists on the Crowdin server. Possibly the
intent was to verify that the path being represented is a valid file
path. Regardless, the existence check was removed to allow retrieval of
a translated file without the source file being present on the client
system.